+++ packages/@headlessui-vue/src/components/menu/menu.test.tsx
@@ -1,6 +1,7 @@
 import { defineComponent, h, nextTick, ref, watch } from 'vue'
 import { render } from '../../test-utils/vue-testing-library'
 import { Menu, MenuButton, MenuItems, MenuItem } from './menu'
+import { TransitionChild } from '../transitions/transition'
 import { suppressConsoleLogs } from '../../test-utils/suppress-console-logs'
 import {
   MenuState,
@@ -40,6 +41,16 @@ beforeAll(() => {
 })
 
 afterAll(() => jest.restoreAllMocks())
+
+function nextFrame() {
+  return new Promise(resolve => {
+    requestAnimationFrame(() => {
+      requestAnimationFrame(() => {
+        resolve()
+      })
+    })
+  })
+}
 
 function renderTemplate(input: string | Partial<Parameters<typeof defineComponent>[0]>) {
   let defaultComponents = { Menu, MenuButton, MenuItems, MenuItem }
@@ -596,7 +607,6 @@ describe('Rendering', () => {
                 '  - id',
                 '  - role',
                 '  - tabIndex',
-                '  - class',
                 '  - aria-disabled',
                 '  - onClick',
                 '  - onFocus',
@@ -622,60 +632,6 @@ describe('Rendering', () => {
 })
 
 describe('Rendering composition', () => {
-  it('should be possible to conditionally render classNames (aka className can be a function?!)', async () => {
-    renderTemplate(jsx`
-      <Menu>
-        <MenuButton>Trigger</MenuButton>
-        <MenuItems>
-          <MenuItem as="a" :className="JSON.stringify">Item A</MenuItem>
-          <MenuItem as="a" disabled :className="JSON.stringify">Item B</MenuItem>
-          <MenuItem as="a" class="no-special-treatment">Item C</MenuItem>
-        </MenuItems>
-      </Menu>
-    `)
-
-    assertMenuButton({
-      state: MenuState.InvisibleUnmounted,
-      attributes: { id: 'headlessui-menu-button-1' },
-    })
-    assertMenu({ state: MenuState.InvisibleUnmounted })
-
-    // Open menu
-    await click(getMenuButton())
-
-    let items = getMenuItems()
-
-    // Verify correct classNames
-    expect('' + items[0].classList).toEqual(JSON.stringify({ active: false, disabled: false }))
-    expect('' + items[1].classList).toEqual(JSON.stringify({ active: false, disabled: true }))
-    expect('' + items[2].classList).toEqual('no-special-treatment')
-
-    // Double check that nothing is active
-    assertNoActiveMenuItem()
-
-    // Make the first item active
-    await press(Keys.ArrowDown)
-
-    // Verify the classNames
-    expect('' + items[0].classList).toEqual(JSON.stringify({ active: true, disabled: false }))
-    expect('' + items[1].classList).toEqual(JSON.stringify({ active: false, disabled: true }))
-    expect('' + items[2].classList).toEqual('no-special-treatment')
-
-    // Double check that the first item is the active one
-    assertMenuLinkedWithMenuItem(items[0])
-
-    // Let's go down, this should go to the third item since the second item is disabled!
-    await press(Keys.ArrowDown)
-
-    // Verify the classNames
-    expect('' + items[0].classList).toEqual(JSON.stringify({ active: false, disabled: false }))
-    expect('' + items[1].classList).toEqual(JSON.stringify({ active: false, disabled: true }))
-    expect('' + items[2].classList).toEqual('no-special-treatment')
-
-    // Double check that the last item is the active one
-    assertMenuLinkedWithMenuItem(items[2])
-  })
-
   it(
     'should be possible to swap the menu item with a button for example',
     suppressConsoleLogs(async () => {
@@ -723,22 +679,22 @@ describe('Rendering composition', () => {
         template: jsx`
           <Menu>
             <MenuButton>Trigger</MenuButton>
-            <div className="outer">
+            <div class="outer">
               <MenuItems>
-                <div className="py-1 inner">
+                <div class="py-1 inner">
                   <MenuItem as="button">Item A</MenuItem>
                   <MenuItem as="button">Item B</MenuItem>
                 </div>
-                <div className="py-1 inner">
+                <div class="py-1 inner">
                   <MenuItem as="button">Item C</MenuItem>
                   <MenuItem>
                     <div>
-                      <div className="outer">Item D</div>
+                      <div class="outer">Item D</div>
                     </div>
                   </MenuItem>
                 </div>
-                <div className="py-1 inner">
-                  <form className="inner">
+                <div class="py-1 inner">
+                  <form class="inner">
                     <MenuItem as="button">Item E</MenuItem>
                   </form>
                 </div>
@@ -880,6 +836,56 @@ describe('Composition', () => {
       expect(readFn).toHaveBeenNthCalledWith(3, State.Open)
     })
   )
+
+  it('should be possible to render a TransitionChild that inherits state from the Menu', async () => {
+    let readFn = jest.fn()
+    renderTemplate({
+      components: { TransitionChild },
+      template: jsx`
+        <Menu>
+          <MenuButton>Trigger</MenuButton>
+          <TransitionChild
+            as="template"
+            @beforeEnter="readFn('enter')"
+            @beforeLeave="readFn('leave')"
+          >
+            <MenuItems>
+              <MenuItem as="button">I am a button</MenuItem>
+            </MenuItems>
+          </TransitionChild>
+        </Menu>
+      `,
+      setup() {
+        return { readFn }
+      },
+    })
+
+    // Verify the Menu is hidden
+    assertMenu({ state: MenuState.InvisibleUnmounted })
+
+    // Let's toggle the Menu
+    await click(getMenuButton())
+
+    // Verify that our transition fired
+    expect(readFn).toHaveBeenCalledTimes(1)
+    expect(readFn).toHaveBeenNthCalledWith(1, 'enter')
+
+    // Verify the Menu is visible
+    assertMenu({ state: MenuState.Visible })
+
+    // Let's toggle the Menu
+    await click(getMenuButton())
+
+    // Verify that our transition fired
+    expect(readFn).toHaveBeenCalledTimes(2)
+    expect(readFn).toHaveBeenNthCalledWith(2, 'leave')
+
+    // Wait for the transitions to finish
+    await nextFrame()
+
+    // Verify the Menu is hidden
+    assertMenu({ state: MenuState.InvisibleUnmounted })
+  })
 })
 
 describe('Keyboard interactions', () => {